+++ src/main/java/domain/GameManager.java
@@ -1,5 +1,10 @@
 package domain;
 
+import domain.card.CardDeck;
+import domain.participant.Participant;
+import domain.participant.Participants;
+import domain.participant.Player;
+import domain.result.BlackjackResult;
 import dto.ResultDto;
 import java.util.Collections;
 import java.util.EnumMap;
