+++ src/main/java/domain/card/TrumpShape.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public enum TrumpShape {
