+++ src/main/java/domain/participant/Participant.java
@@ -1,7 +1,9 @@
-package domain;
+package domain.participant;
+
+import domain.card.Card;
 
 public interface Participant {
-    
+
     void receive(Card card);
 
     int getScore();
