+++ src/test/java/domain/GameManagerTest.java
@@ -3,6 +3,14 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import domain.card.Card;
+import domain.card.CardDeck;
+import domain.card.TrumpNumber;
+import domain.card.TrumpShape;
+import domain.participant.Dealer;
+import domain.participant.Participants;
+import domain.participant.Player;
+import domain.result.BlackjackResult;
 import dto.ResultDto;
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
