+++ src/test/java/domain/card/CardsTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
