+++ src/test/java/domain/participant/ParticipantsTest.java
@@ -1,9 +1,10 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.card.CardDeck;
 import java.util.List;
 import org.junit.jupiter.api.Test;
