+++ src/test/java/domain/result/BlackjackResultTest.java
@@ -1,7 +1,12 @@
-package domain;
+package domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.Card;
+import domain.card.TrumpNumber;
+import domain.card.TrumpShape;
+import domain.participant.Dealer;
+import domain.participant.Player;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
